requirements.txt
numpy==1.20.2
opencv_python==4.5.5.64
Pillow==9.2.0
streamlit==1.12.2
